requirements.txt
simpleaudio
gtts 
kafka-python 
pydub